Added in player health and damage to player
Created MechHealth script on the player object and created a healthbar on the UI with a fillable image.
Added functionality in PlayerStatsDisplay to update the healthbar and created a delegate + event listener
Added functionality in MechHealth to manipulate the health up and down, as well as delegate/event listener for invoking those methods
Add code to DemonClaw.cs which invokes the MechHealthDown event
Player is now hurt by demons!