Dialogue, textures and props
Adding door with lock script & key script
Added in KeyRing script which has bools for red, blue and yellow keys.
Has a method receiving an int (keycolour) and checking if the corresponding bool is true
has a method receiving an int (keycolour) for picking up a key which sets the corresponding bool to true
Added in ExitDoor script with an enum for the door colour and an ontrigger method checking if the object is the player then invoking the KeyCheckEvent on the KeyRing to check if the players has the key
Added Key script with an enum showing what kind of key it is and an ontrigger method which will invoke the AddKeyEvent on the keyring to add that colour key to the key ring
Key also disables itself once picked up.
Added an animation to the door so that it opens when the player interacts with it while holding the key.
Dialogue/info HUD
Created UI panel called DialogueHUD
Added DialogueDisplay script with delegate for toggling the dialogue display on/off and displaying text.
Add a dialogue prompt showing when key is collected
Added dialologue prompt for approaching door with no key and approaching door with key

Added functionality to the DemonHealthBar to toggle it on only after a Demon has taken damage.Then realised that there are multiple instances of demon health bar and I'm stupid so it will have to be a direct reference lol

Added in primited objects as walls with nev mesh obstacle components to carve out the navmesh
Added basic materials to walls and floor and door
Added an object to be used as the 'you win' trigger
made bullets detect walls and delyeet themselves
Made the demon spawner a prefab and placed spawners into each room